remove unused code and restructure configuration
