+++ client/src/ClassSelectionMenu/ClassSelectionMenu.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, fireEvent , screen} from '@testing-library/react';
+import { render, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom';
 import ClassSelectionMenu from './index';
