+++ client/src/ConfigureImageClassification/ConfigureImageClassification.test.js
@@ -97,6 +97,5 @@ describe('ConfigureImageClassification', () => {
 
         // Simulate completing the form
         fireEvent.click(screen.getByTestId('complete-button'));
-        // Add assertions to verify completion behavior based on YourComponent logic
     });
 });
