+++ client/src/SetupPage/index.jsx
@@ -19,7 +19,7 @@ import { useSettings } from "../SettingsProvider";
 import { useTranslation } from "react-i18next"
 import { Info } from '@mui/icons-material';
 import NoteSection from "../NoteSection";
-import config from '../../config.js'; //importing demo site config from client/config.js
+import config from '../config.js';
 
 const Container = styled("div")({
   marginTop: "2rem",
