+++ client/src/utils/send-data-to-server.js
@@ -1,5 +1,5 @@
 import axios from 'axios'
-import config from '../../config.js'
+import config from '../config.js'
 
 export const getImageData = (activeImage) => {
   let imageData = {}
