+++ client/src/workspace/DownloadButton/DownloadButton.test.js
@@ -10,7 +10,7 @@ jest.mock('../../SnackbarContext/index.jsx', () => ({
   }),
 }));
 
-jest.mock('../../../config.js', () => ({
+jest.mock('../../config.js', () => ({
     DEMO_SITE_URL: "https://annotate-docs.dwaste.live/",
     VITE_SERVER_URL: "http://localhost:5000",
   }));
