+++ client/src/workspace/Header/Header.test.js
@@ -3,7 +3,7 @@ import { render, screen, fireEvent } from "@testing-library/react";
 import Header from "./index";
 import '@testing-library/jest-dom';
 
-jest.mock('../../../config.js', () => ({
+jest.mock('../../config.js', () => ({
     DEMO_SITE_URL: "https://annotate-docs.dwaste.live/",
     VITE_SERVER_URL: "http://localhost:5000",
   }));
